Refactor ShipmentSending Schedules: make orthogonal to Locations
I have moved the burden of sending shipment schedules to outside of the
confines of the locations. Sending shipments is really orthogonal to the
definition of locations. Sending a shipment from one location to another
is best defined outside the locations, and instead in a third class of actors.
These actors will keep track of time, and send things along appropriately. This
is the easiest in terms of extensibility, etc.